use more precise timer `timeit.default_timer` for log_durations
Instead of using `time.time`, this uses \`timeit.default_timer()\`.
`default_timer` uses `time.perf_counter` which is monotonic,
is not user-adjustable, and is more precise.

However, `time.perf_counter` is only available since Python>3.3,
and hence is unavailable in Python 2.7. In that version, it defaults
to `time.clock()` in Windows and `time.time()` in other platforms.